# docs/openapi.yaml
openapi: 3.0.3
info:
  title: SmartStore SaaS API
  description: |
    A comprehensive REST API for managing e-commerce operations including products, orders, customers, and analytics.
    
    ## Authentication
    All API endpoints require authentication using NextAuth.js with JWT tokens.
    
    ## Rate Limiting
    API endpoints are rate limited to prevent abuse:
    - Authentication endpoints: 10 requests per minute
    - CRUD operations: 100 requests per 15 minutes
    - Analytics endpoints: 50 requests per 15 minutes
    - Monitoring endpoints: 200 requests per 15 minutes
    
    ## Support
    For API support and questions:
    - Documentation: https://docs.smartstore-saas.com
    - Support Email: support@smartstore-saas.com
    - Status Page: https://status.smartstore-saas.com
  version: 1.2.0
  contact:
    name: SmartStore SaaS Support
    email: support@smartstore-saas.com
    url: https://docs.smartstore-saas.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://smart-store-saas-demo.vercel.app/api
    description: Production server
  - url: http://localhost:3000/api
    description: Development server
security:
  - BearerAuth: []
paths:
...
  /orders:
    get:
      tags:
        - Orders
      summary: Get Orders
      description: Retrieve a paginated list of orders with optional filtering
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 10
        - name: status
          in: query
          description: Filter by order status
          schema:
            type: string
            enum: [pending, processing, shipped, delivered, cancelled]
        - name: customerId
          in: query
          description: Filter by customer ID
          schema:
            type: string
        - name: dateFrom
          in: query
          description: Start date (ISO format)
          schema:
            type: string
            format: date-time
        - name: dateTo
          in: query
          description: End date (ISO format)
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: Orders retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Order'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
        '401':
          $ref: '#/components/responses/UnauthorizedError'

    post:
      tags:
        - Orders
      summary: Create Order
      description: Create a new order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrderCreate'
      responses:
        '201':
          description: Order created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/Order'
                  message:
                    type: string
                    example: Order created successfully
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /customers:
    get:
      tags:
        - Customers
      summary: Get Customers
      description: Retrieve a paginated list of customers with optional filtering
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 10
        - name: search
          in: query
          description: Search term for name, email, or phone
          schema:
            type: string
        - name: segment
          in: query
          description: Filter by customer segment
          schema:
            type: string
      responses:
        '200':
          description: Customers retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Customer'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
        '401':
          $ref: '#/components/responses/UnauthorizedError'

    post:
      tags:
        - Customers
      summary: Create Customer
      description: Create a new customer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CustomerCreate'
      responses:
        '201':
          description: Customer created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/Customer'
                  message:
                    type: string
                    example: Customer created successfully
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /analytics/dashboard:
    get:
      tags:
        - Analytics
      summary: Get Dashboard Statistics
      description: Retrieve dashboard statistics including counts and recent activity
      responses:
        '200':
          description: Dashboard statistics retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/DashboardStats'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Order:
      type: object
      properties:
        id:
          type: string
          example: order-123
        orderNumber:
          type: string
          example: ORD-001
        status:
          type: string
          enum: [pending, processing, shipped, delivered, cancelled]
          example: pending
        total:
          type: number
          format: decimal
          example: 99.98
        customerId:
          type: string
          example: customer-123
        createdAt:
          type: string
          format: date-time
          example: 2024-01-01T00:00:00Z
        updatedAt:
          type: string
          format: date-time
          example: 2024-01-01T00:00:00Z
        customer:
          type: object
          properties:
            id:
              type: string
              example: customer-123
            name:
              type: string
              example: John Doe
            email:
              type: string
              example: john@example.com
        orderItems:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
                example: item-123
              quantity:
                type: integer
                example: 2
              price:
                type: number
                example: 29.99
              product:
                type: object
                properties:
                  name:
                    type: string
                    example: Sample Product
                  sku:
                    type: string
                    example: SKU-001
    OrderCreate:
      type: object
      required:
        - customerId
        - items
      properties:
        customerId:
          type: string
          example: customer-123
        items:
          type: array
          items:
            type: object
            required:
              - productId
              - quantity
              - price
            properties:
              productId:
                type: string
                example: product-123
              quantity:
                type: integer
                minimum: 1
                example: 2
              price:
                type: number
                minimum: 0.01
                example: 29.99
        shippingAddress:
          type: object
          properties:
            street:
              type: string
              example: 123 Main St
            city:
              type: string
              example: Anytown
            state:
              type: string
              example: CA
            zipCode:
              type: string
              example: 12345
            country:
              type: string
              example: US
    Customer:
      type: object
      properties:
        id:
          type: string
          example: customer-123
        name:
          type: string
          example: John Doe
        email:
          type: string
          format: email
          example: john@example.com
        phone:
          type: string
          example: +1234567890
        address:
          type: object
          properties:
            street:
              type: string
              example: 123 Main St
            city:
              type: string
              example: Anytown
            state:
              type: string
              example: CA
            zipCode:
              type: string
              example: 12345
        createdAt:
          type: string
          format: date-time
          example: 2024-01-01T00:00:00Z
        updatedAt:
          type: string
          format: date-time
          example: 2024-01-01T00:00:00Z
        _count:
          type: object
          properties:
            orders:
              type: integer
              example: 5
    CustomerCreate:
      type: object
      required:
        - name
        - email
      properties:
        name:
          type: string
          minLength: 1
          example: Jane Doe
        email:
          type: string
          format: email
          example: jane@example.com
        phone:
          type: string
          example: +1234567891
        address:
          type: object
          properties:
            street:
              type: string
              example: 456 Oak Ave
            city:
              type: string
              example: Another City
            state:
              type: string
              example: NY
            zipCode:
              type: string
              example: 67890
    Pagination:
      type: object
      properties:
        page:
          type: integer
          example: 1
        limit:
          type: integer
          example: 10
        total:
          type: integer
          example: 100
        totalPages:
          type: integer
          example: 10
    DashboardStats:
      type: object
      properties:
        products:
          type: integer
          example: 150
        orders:
          type: integer
          example: 500
        customers:
          type: integer
          example: 102
        revenue:
          type: number
          format: decimal
          example: 25000.50
        recentOrders:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
                example: order-123
              orderNumber:
                type: string
                example: ORD-001
              total:
                type: number
                example: 99.98
              status:
                type: string
                example: completed
              createdAt:
                type: string
                format: date-time
        topProducts:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
                example: product-123
              name:
                type: string
                example: Best Seller
              sales:
                type: integer
                example: 50
              revenue:
                type: number
                example: 1499.50
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
          example: Error type
        message:
          type: string
          example: Human-readable error message
        code:
          type: string
          example: ERROR_CODE
        timestamp:
          type: string
          format: date-time
          example: 2024-01-01T00:00:00Z
        path:
          type: string
          example: /api/endpoint
        requestId:
          type: string
          example: req_1234567890_abcdef
    ValidationError:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
          example: Validation failed
        details:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
                example: email
              message:
                type: string
                example: Email is required
              code:
                type: string
                example: required
  responses:
    UnauthorizedError:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    ValidationError:
      description: Validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationError'
...